Use sequence builder helper in pipeline
